Move all middleware from a separate module to server/middleware.go This resolves some potential import cycles
