--- server/middleware/doc.go
@@ -1,4 +1,0 @@
-// Package middleware defines Apollo-specific middleware.
-// All middleware is compatible with Go's net/http module so you could use
-// it standalone, without using the rest of the Apollo library.
-package middleware
